DialogAddInstance: Input validation. More cosmetic improvements to code

+++ src/gui/FormMain.java
@@ -51,7 +51,7 @@ public class FormMain extends JFrame {
 
 	/* Parameters to be tuned */
 //	long maxIterations = 100000L;
-    long maxIterations = Integer.MAX_VALUE;
+    int maxIterations = Integer.MAX_VALUE;
 	//			long maxIterations = 30;
 	int numberOfNeighbors = 500;
 	int sleepDuration = 0;	// in ms
